[Modal] splitted components
